Fix Slug Firing Behavior
Reported by Eric on the HDDiscord. The Altis, for a while, did not fire the proper bullet actors whenever slugs were loaded into the weapon (this is due to an oversight regarding case statements in the new fire function). I also added a bullet-wad to the slugs when fired, since apparently this was absent in the earlier revision.